+++ components/HomePage/home.jsx
@@ -1,10 +1,10 @@
 import Button from "../Button/button";
 import Triangle from "../Triangle/triangle";
-import Information1 from "../Website Information/what is spaceslounge/information-1";
+import Information from "../WebsiteInformation/WhatIsSpaceslounge/information";
 
 import { motion as m } from "framer-motion";
 
-const home = () => {
+const Home = () => {
   return (
     <div>
       <m.header exit={{ opacity: 0 }} transition={{ duration: 0.3 }}>
@@ -31,9 +31,9 @@ const home = () => {
         </div>
       </m.header>
 
-      <Information1 />
+      <Information />
     </div>
   );
 };
 
-export default home;
+export default Home;
