+++ components/WebsiteInformation/WhatIsSpaceslounge/information.scss
@@ -1,4 +1,4 @@
-.information_1 {
+.information {
   padding: 5em 0;
   color: var(--text-color);
 }
